Move button in LoginWindow to app_buttons_login.py
